requirements.txt
aiogram>=3.0.0
aiosqlite
python-dotenv
google-genai
apscheduler
groq
